modularize refactor fast api server
